003-stack_cloud_functions-main/cf_parquet_to_bigquery/requirements.txt
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-logging
fsspec
gcsfs
pyarrow
pandas_gbq